# src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "ecommerce-spring-reactjs"
servers:
  - url: "http://localhost:8080"
  - url: "https://perfume-websore-api.herokuapp.com"
...
paths:
  /api/v1/admin/add:
    post:
      tags:
        - "AdminController"
      summary: "Add perfume to database"
      operationId: "addPerfume"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                perfumeTitle:
                  type: string
                perfumer:
                  type: string
                year:
                  type: string
                country:
                  type: string
                perfumeGender:
                  type: string
                fragranceTopNotes:
                  type: string
                fragranceMiddleNotes:
                  type: string
                fragranceBaseNotes:
                  type: string
                description:
                  type: string
                filename:
                  type: string
                price:
                  type: string
                volume:
                  type: string
                type:
                  type: string
              required:
                - perfumeTitle
                - perfumer
                - year
                - country
                - perfumeGender
                - fragranceTopNotes
                - fragranceMiddleNotes
                - fragranceBaseNotes
                - price
                - volume
                - type
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/FullPerfumeResponse"
              examples:
                PerfumeResponse:
                  $ref: "#/components/schemas/FullPerfumeResponse"
        "400":
          description: "Bad request"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeError"
              examples:
                PerfumeError:
                  $ref: "#/components/schemas/PerfumeError"
      security:
        - token: []
  /api/v1/admin/edit:
    post:
      tags:
        - "AdminController"
      summary: "Update perfume"
      operationId: "updatePerfume"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                perfumeTitle:
                  type: string
                perfumer:
                  type: string
                year:
                  type: string
                country:
                  type: string
                perfumeGender:
                  type: string
                fragranceTopNotes:
                  type: string
                fragranceMiddleNotes:
                  type: string
                fragranceBaseNotes:
                  type: string
                description:
                  type: string
                filename:
                  type: string
                price:
                  type: string
                volume:
                  type: string
                type:
                  type: string
              required:
                - perfumeTitle
                - perfumer
                - year
                - country
                - perfumeGender
                - fragranceTopNotes
                - fragranceMiddleNotes
                - fragranceBaseNotes
                - price
                - volume
                - type
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/FullPerfumeResponse"
              examples:
                PerfumeResponse:
                  $ref: "#/components/schemas/FullPerfumeResponse"
        "400":
          description: "Bad request"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeError"
              examples:
                PerfumeError:
                  $ref: "#/components/schemas/PerfumeError"
      security:
        - token: []
  /api/v1/admin/delete/{id}:
    delete:
      tags:
        - "AdminController"
      summary: "Delete perfume"
      operationId: "deletePerfume"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Perfume deleted successfully"
        "404":
          description: "Not Found"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Perfume not found."
      security:
        - token: []
  /api/v1/admin/orders:
    get:
      tags:
        - "AdminController"
      summary: "Get all orders from database"
      operationId: "getAllOrders"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/order/{email}:
    get:
      tags:
        - "AdminController"
      summary: "Get user orders by email"
      operationId: "getUserOrdersByEmail"
      parameters:
        - name: "email"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/order/delete/{id}:
    delete:
      tags:
        - "AdminController"
      summary: "Delete order"
      operationId: "deleteOrder"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Order deleted successfully"
        "404":
          description: "Not Found"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Order not found."
      security:
        - token: []
...
  /api/v1/admin/user/all:
    get:
      tags:
        - "AdminController"
      summary: "Get users from database"
      operationId: "getAllUsers"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BaseUserResponse"
      security:
        - token: []
  /api/v1/admin/graphql/user:
    post:
      tags:
        - "AdminController"
      summary: "Get user by GraphQL query"
      operationId: "getUserByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ user(id: 2) { id email firstName lastName city address phoneNumber postIndex } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  email:
                    type: "string"
                  firstName:
                    type: "string"
                  lastName:
                    type: "string"
                  city:
                    type: "string"
                  address:
                    type: "string"
                  phoneNumber:
                    type: "string"
                  postIndex:
                    type: "string"
                example:
                  id: 2
                  email: "test123@test.com"
                  firstName: "John"
                  lastName: "Doe"
                  city: "New York"
                  address: "Wall Street 1"
                  phoneNumber: "1234567890"
                  postIndex: "1234567890"
      security:
        - token: []
...
  /api/v1/admin/graphql/orders:
    post:
      tags:
        - "AdminController"
      summary: "Get all orders by GraphQL query"
      operationId: "getAllOrdersQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ orders { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/admin/graphql/order:
    post:
      tags:
        - "AdminController"
      summary: "Get orders by user Email GraphQL query"
      operationId: "getUserOrdersByEmailQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
...
  /api/v1/auth/forgot/{email}:
    get:
      tags:
        - "AuthenticationController"
      summary: "Send password reset code to users email"
      operationId: "forgotPassword"
      parameters:
        - name: "email"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Reset password code is send to your E-mail"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Email not found"
...
  /api/v1/auth/reset:
    post:
      tags:
        - "AuthenticationController"
      summary: "Reset password"
      operationId: "passwordReset"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test@test.com"
                password:
                  type: string
                  example: "1234567890"
                password2:
                  type: string
                  example: "1234567890"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                PasswordMatchError:
                  summary: "Password Match Error"
                  value:
                    passwordError: "Passwords do not match."
                PasswordConfirmationError:
                  summary: "Password Confirmation Error"
                  value:
                    password2Error: "Password confirmation cannot be empty."
  /api/v1/auth/edit/password:
    put:
      tags:
        - "AuthenticationController"
      summary: "Change user password"
      operationId: "updateUserPassword"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: "string"
                  example: "1234567890"
                password2:
                  type: "string"
                  example: "1234567890"
              required:
                - password
                - password2
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                BindingResultError:
                  summary: "Binding Result Errors"
                  value:
                    passwordError: "The password must be between 6 and 16 characters long"
                    password2Error: "The password confirmation must be between 6 and 16 characters long"
                MatchPasswordError:
                  summary: "Match Password Error"
                  value:
                    data: "Passwords do not match."
      security:
        - token: []
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    FullPerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        perfumeRating:
          type: "integer"
          format: "int32"
          example: 5
        year:
          type: "integer"
          format: "int32"
          example: 2015
        country:
          type: "string"
          example: "France"
        perfumeGender:
          type: "string"
          example: "male"
        fragranceTopNotes:
          type: "string"
          example: "Sichuan Pepper"
        fragranceMiddleNotes:
          type: "string"
          example: "Bergamot"
        fragranceBaseNotes:
          type: "string"
          example: "Ambroxan"
        description:
          type: "string"
          example: "Perfume description"
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        price:
          type: "integer"
          format: "int32"
          example: 62
        volume:
          type: "string"
          example: "100"
        type:
          type: "string"
          example: "Eau de parfum"
        file:
          type: "string"
          format: binary
...
    PerfumeError:
      type: "object"
      properties:
        perfumeTitleError:
          type: "string"
          example: "Fill in the input field"
        perfumerError:
          type: "string"
          example: "Fill in the input field"
        yearError:
          type: "string"
          example: "Fill in the input field"
        countryError:
          type: "string"
          example: "Fill in the input field"
        typeError:
          type: "string"
          example: "Fill in the input field"
        volumeError:
          type: "string"
          example: "Fill in the input field"
        perfumeGenderError:
          type: "string"
          example: "Fill in the input field"
        fragranceTopNotesError:
          type: "string"
          example: "Fill in the input field"
        fragranceMiddleNotesError:
          type: "string"
          example: "Fill in the input field"
        fragranceBaseNotesError:
          type: "string"
          example: "Fill in the input field"
        priceError:
          type: "string"
          example: "Fill in the input field"
    OrderResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        totalPrice:
          type: "number"
          format: "double"
          example: 100
        date:
          type: "string"
          example: "2021-04-07"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        email:
          type: "string"
          example: "test123@test.com"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "integer"
          format: "int32"
          example: 123456
...
    BaseUserResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 2
        email:
          type: "string"
          example: "test123@test.com"
        firstName:
          type: "string"
          example: "John"
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
        provider:
          type: "string"
          example: "LOCAL"
...